Fixed cast bug

+++ src/Venflow/Venflow/UInt64Handler.cs
@@ -5,6 +5,6 @@ namespace Venflow
     internal class UInt64Handler : IParameterTypeHandler
     {
         NpgsqlParameter IParameterTypeHandler.Handle(string name, object val)
-            => new NpgsqlParameter<long>(name, unchecked((long)val + long.MinValue));
+            => new NpgsqlParameter<long>(name, unchecked((long)(ulong)val + long.MinValue));
     }
 }
